+++ src/test/resources/allTests.xml
@@ -5,7 +5,6 @@
         <classes>
             <class name="TestsForMan"/>
             <class name="TestsForWoman"/>
-            <class name="TestsForGettersAndSetters"/>
         </classes>
     </test>
 </suite>
